algo/src/main/java/com/hartwig/actin/algo/evaluation/treatment/SecondMalignanciesHaveBeenCuredRecently.java
package com.hartwig.actin.algo.evaluation.treatment;

import com.hartwig.actin.PatientRecord;
import com.hartwig.actin.algo.datamodel.Evaluation;
import com.hartwig.actin.algo.datamodel.EvaluationResult;
import com.hartwig.actin.algo.evaluation.EvaluationFactory;
import com.hartwig.actin.algo.evaluation.EvaluationFunction;
import com.hartwig.actin.clinical.datamodel.PriorSecondPrimary;

import org.jetbrains.annotations.NotNull;

public class SecondMalignanciesHaveBeenCuredRecently implements EvaluationFunction {

    SecondMalignanciesHaveBeenCuredRecently() {
    }

    @NotNull
    @Override
    public Evaluation evaluate(@NotNull PatientRecord record) {
        for (PriorSecondPrimary priorSecondPrimary : record.clinical().priorSecondPrimaries()) {
            if (priorSecondPrimary.isActive()) {
                return EvaluationFactory.create(EvaluationResult.FAIL);
            }
        }
        return EvaluationFactory.create(EvaluationResult.PASS);
    }
}
